added blogs controller
